# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/memberships/{id}": {
      "patch": {
        "operationId": "AdminController_updateMembership",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/invitations": {
      "post": {
        "operationId": "AdminController_invite",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "get": {
        "operationId": "AdminController_listInvitations",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/invitations/{id}/revoke": {
      "post": {
        "operationId": "AdminController_revokeInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lodgiva API",
    "description": "Hotel management platform API. All protected routes require a Bearer access token; the tenant is always derived from the token and never from the request body.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}